backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=21.2
dj-database-url>=2.1
Faker>=24.0
django-redis>=5.4
drf-spectacular>=0.27
djangorestframework-simplejwt>=5.3
pandas>=2.0
celery>=5.3


